Minecraft Game - Python Project
Minecraft is the most popular game on the internet and you can play it on windows android and play station. By using the following code you are going to make the basic clone of Minecraft. You are going to make it in python using the Ursina engine. It is an open-source game engine for 3d game development in python. You can use Ursina only in Linux and windows it doesn’t run in Mac. You are going to play with boxes in this game. you can create and destroy the blocks and in this game, you are going to make a house using these blocks.